pyproject.toml
[tool.poetry]
name = "course-7-coursework"
version = "0.1.0"
description = ""
authors = ["Vladimir Titov <tvv-vg@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
celery = "^5.4.0"
redis = "^5.0.6"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-celery-beat = "^2.6.0"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
mypy = "^1.10.0"


[tool.poetry.group.test.dependencies]
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
